@@ -8,13 +8,13 @@ Rails.application.config.middleware.insert_before 0, Rack::Cors, debug: true, lo
     origins '*'
 
     resource '/cors',
-      :headers => :any,
-      :methods => [:post],
-      :max_age => 0
+             headers: :any,
+             methods: [:post],
+             max_age: 0
 
     resource '*',
-      :headers => :any,
-      :methods => [:get, :post, :delete, :put, :patch, :options, :head],
-      :max_age => 0
+             headers: :any,
+             methods: [:get, :post, :delete, :put, :patch, :options, :head],
+             max_age: 0
   end
 end
