@@ -2,12 +2,12 @@ module SerializationConcern
   extend ActiveSupport::Concern
 
   def render_one(object, serializer_class: nil, status: :ok, options: {})
-    return render_no_content unless object.present?
+    return render_no_content if object.blank?
 
     serializer = serializer_class&.new(**options) ||
                  guess_serializer(object.class).new(**options)
     json = serializer.serialize_to_json(object)
-    render json: json, status: status
+    render json:, status:
   end
 
   def render_many(resources, serializer_class: nil, status: :ok, options: {})
@@ -19,7 +19,7 @@ module SerializationConcern
     )
 
     json = serializer.to_json
-    render json: json, status: status
+    render json:, status:
   end
 
   def render_no_content
