@@ -1,5 +1,5 @@
 class UnprocessableEntityError < ControllerError
   def initialize(message: 'Unprocessable Entity', status: 422)
-    super(status: status, message: message)
+    super(status:, message:)
   end
 end
