@@ -30,15 +30,15 @@ module DemocratifyAPI
     # in config/environments, which are processed later.
     #
     # config.time_zone = "Central Time (US & Canada)"
-    config.eager_load_paths << Rails.root.join('app', 'poros', 'integrations')
-    config.eager_load_paths << Rails.root.join('app', 'models', 'concerns')
+    config.eager_load_paths << Rails.root.join('app/poros/integrations')
+    config.eager_load_paths << Rails.root.join('app/models/concerns')
 
     # Only loads a smaller set of middleware suitable for API only apps.
     # Middleware like session, flash, cookies can be added back manually.
     # Skip views, helpers and assets when generating a new resource.
     config.api_only = true
 
-    config.autoload_paths << Rails.root.join('app', 'poros', 'integrations')
-    config.autoload_paths << Rails.root.join('app', 'models', 'concerns')
+    config.autoload_paths << Rails.root.join('app/poros/integrations')
+    config.autoload_paths << Rails.root.join('app/models/concerns')
   end
 end
