@@ -60,7 +60,7 @@ ActiveRecord::Schema.define(version: 2022_01_03_142044) do
     t.datetime "updated_at", precision: 6, null: false
     t.bigint "spotify_playlist_id", null: false
     t.bigint "spotify_song_id", null: false
-    t.index ["spotify_playlist_id", "index"], name: "index_spotify_playlist_songs_on_spotify_playlist_id_and_index"
+    t.index %w[spotify_playlist_id index], name: "index_spotify_playlist_songs_on_spotify_playlist_id_and_index"
     t.index ["spotify_playlist_id"], name: "index_spotify_playlist_songs_on_spotify_playlist_id"
     t.index ["spotify_song_id"], name: "index_spotify_playlist_songs_on_spotify_song_id"
   end
@@ -153,7 +153,7 @@ ActiveRecord::Schema.define(version: 2022_01_03_142044) do
     t.datetime "updated_at", precision: 6, null: false
     t.bigint "ongoing_playlist_id", null: false
     t.bigint "account_id", null: false
-    t.index ["account_id", "ongoing_playlist_id"], name: "index_votations_on_account_id_and_ongoing_playlist_id"
+    t.index %w[account_id ongoing_playlist_id], name: "index_votations_on_account_id_and_ongoing_playlist_id"
     t.index ["account_id"], name: "index_votations_on_account_id"
     t.index ["ongoing_playlist_id"], name: "index_votations_on_ongoing_playlist_id"
   end
