@@ -29,6 +29,6 @@
 #
 FactoryBot.define do
   factory :votation do
-    
+    pending
   end
 end
