@@ -17,6 +17,6 @@
 #
 require 'rails_helper'
 
-RSpec.describe Account, type: :model do
+RSpec.describe Account do
   pending "add some examples to (or delete) #{__FILE__}"
 end
