@@ -24,6 +24,6 @@
 #
 require 'rails_helper'
 
-RSpec.describe SpotifyDevice, type: :model do
+RSpec.describe SpotifyDevice do
   pending "add some examples to (or delete) #{__FILE__}"
 end
