@@ -25,6 +25,6 @@
 #
 require 'rails_helper'
 
-RSpec.describe SpotifySong, type: :model do
+RSpec.describe SpotifySong do
   pending "add some examples to (or delete) #{__FILE__}"
 end
