@@ -16,6 +16,6 @@
 #
 require 'rails_helper'
 
-RSpec.describe User, type: :model do
+RSpec.describe User do
   pending "add some examples to (or delete) #{__FILE__}"
 end
