@@ -27,6 +27,6 @@
 #
 require 'rails_helper'
 
-RSpec.describe VotationCandidate, type: :model do
+RSpec.describe VotationCandidate do
   pending "add some examples to (or delete) #{__FILE__}"
 end
