@@ -29,6 +29,6 @@
 #
 require 'rails_helper'
 
-RSpec.describe Votation, type: :model do
+RSpec.describe Votation do
   pending "add some examples to (or delete) #{__FILE__}"
 end
