@@ -1,6 +1,6 @@
 require 'rails_helper'
 
-RSpec.describe "SpotifySongs", type: :request do
+RSpec.describe "SpotifySongs" do
   describe "GET /index" do
     pending "add some examples (or delete) #{__FILE__}"
   end
